Count line indentation manually
The language service `getIndentationAtPosition` returned different
values for the following two code snippets:

    // Indentation returned for starting `: 4
    const query = sql("delete-orders")`
        DELETE FROM orders
    `;

    // Indentation returned for starting `: 0
    const query = sql`
        DELETE FROM orders
    `;

I couldn't figure out why this is the case. This changes indentation
detection to look at the code in the line and count leading spaces/tabs.
The logic for this has been copied from TypeScript's implementation of
the `getFormattingEditsAfterKeystroke` function.